DISCLAIMER: Read  below.
Hi first off, I DO NOT own the code in this bot, I simply threw it together in a file for this to work. I have edited it for my own server that is all I have done. I do not claim this work as my own. All orginal code go to the owners who deserves the credit. I simply made this bot for others who dont know how to do jack shit, you MUST have NPM installed on a VPS in order to run, the bot 24/7.